requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dicttoxml==1.7.16
Flask==2.3.2
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.3
zipp==3.15.0
